+++ pkg/sstable/writer.go
@@ -46,7 +46,7 @@ func (w *Writer) Write(out io.Writer) (*Meta, error) {
 
 	m := &Meta{
 		Created: w.clock.Now(),
-		Size:    int64(len(magicBytes)),
+		Size:    len(magicBytes),
 	}
 
 	for _, record := range w.records {
@@ -56,7 +56,7 @@ func (w *Writer) Write(out io.Writer) (*Meta, error) {
 		}
 
 		m.Count++
-		m.Size += int64(n)
+		m.Size += n
 
 		if m.MinKey == "" || record.Key < m.MinKey {
 			m.MinKey = record.Key
